feat: add autopin-toggle to constants for toolbar and participants pane
